Push from private host
Moving Protheus back into the public domain.

+++ src/Core/Graphics/Lighting.cpp
@@ -24,11 +24,11 @@ void Lighting::BindLights(Program& program) {
 	}
 
 	program.SetUniform("light_ambient", ambient_color_); 
-	program.SetUniform("light_point_count", light_points_.size());
-	program.SetUniform("light_points", points.data<float>(), light_points_.size() * 8);
-	program.SetUniform("light_directional_count", light_directional_.size());
+	program.SetUniform("light_point_count", (int)light_points_.size());
+	program.SetUniform("light_points", points.data<float>(), (int)light_points_.size() * 8);
+	program.SetUniform("light_directional_count", (int)light_directional_.size());
 	program.SetUniform("light_directional",
-		points.data<float>() + (light_points_.size() * 8), light_directional_.size() * 11); 
+		points.data<float>() + (light_points_.size() * 8), (int)light_directional_.size() * 11);
 }
 
 LightPoint& Lighting::AttachLight(const LightPoint& point) {
